+++ src/biz/bokhorst/xprivacy/UpdateService.java
@@ -128,12 +128,14 @@ public class UpdateService extends IntentService {
 		// Check if something to do
 		boolean work = false;
 		File prefs = new File(Util.getUserDataDirectory(Process.myUid()) + File.separator + "shared_prefs");
-		for (File file : prefs.listFiles())
-			if (!file.getName().startsWith("biz.bokhorst.xprivacy.provider.usage.") && file.getName().endsWith(".xml")
-					&& !new File(file + ".migrated").exists()) {
-				work = true;
-				break;
-			}
+		File[] files = prefs.listFiles();
+		if (files != null)
+			for (File file : files)
+				if (!file.getName().startsWith("biz.bokhorst.xprivacy.provider.usage.")
+						&& file.getName().endsWith(".xml") && !new File(file + ".migrated").exists()) {
+					work = true;
+					break;
+				}
 
 		// Perform migration
 		if (work) {